Additional code clean up
PersonRepository updates
Verified that async and await aren't required for the two methods that
had them because I'm already using ToListAsync.

Can return the new PersonId in the Add method. Confirmed
Context.SaveChanges is updating the PersonID of person.

PersonModelTests updates

CanGetById: Even though the db is empty before the context was adding
people, the ID sequence did not begin with one.  So, updating to default
to using the ID of the first person in people.

CanGetById: Update to the Add in the PersonRepository made it possible
to write the test to be more precise about the data of the person the
test was running against.